removed route-builder-trait, implemented tab-route-builder-trait
