+++ src/Sulu/Bundle/AdminBundle/Admin/Routing/ResourceTabRouteBuilder.php
@@ -13,8 +13,6 @@ namespace Sulu\Bundle\AdminBundle\Admin\Routing;
 
 class ResourceTabRouteBuilder implements ResourceTabRouteBuilderInterface
 {
-    use RouteBuilderTrait;
-
     const VIEW = 'sulu_admin.resource_tabs';
 
     /**
@@ -29,21 +27,23 @@ class ResourceTabRouteBuilder implements ResourceTabRouteBuilderInterface
 
     public function setResourceKey(string $resourceKey): ResourceTabRouteBuilderInterface
     {
-        $this->setResourceKeyToRoute($this->route, $resourceKey);
+        $this->route->setOption('resourceKey', $resourceKey);
 
         return $this;
     }
 
     public function addLocales(array $locales): ResourceTabRouteBuilderInterface
     {
-        $this->addLocalesToRoute($this->route, $locales);
+        $oldLocales = $this->route->getOption('locales');
+        $newLocales = $oldLocales ? array_merge($oldLocales, $locales) : $locales;
+        $this->route->setOption('locales', $newLocales);
 
         return $this;
     }
 
     public function setBackRoute(string $backRoute): ResourceTabRouteBuilderInterface
     {
-        $this->setBackRoute($this->route, $backRoute);
+        $this->route->setOption('backRoute', $backRoute);
 
         return $this;
     }
